Moving to conf system - NumberDisplay
