+++ src/charts/i-number-display-conf.ts
@@ -1,0 +1,6 @@
+import {IBaseMixinConf} from '../base/i-base-mixin-conf';
+import {NumberFormatFn} from '../core/types';
+
+export interface INumberDisplayConf extends IBaseMixinConf {
+    readonly formatNumber?: NumberFormatFn;
+}
